debugged for my laptop cpu

+++ diffusion/fp16_util.py
@@ -17,9 +17,11 @@ def convert_module_to_f16(l):
     Convert primitive modules to float16.
     """
     if isinstance(l, (nn.Conv1d, nn.Conv2d, nn.Conv3d)):
-        l.weight.data = l.weight.data.half()
+        l.weight.data = l.weight.data.float()
+        # l.weight.data = l.weight.data.half()
         if l.bias is not None:
-            l.bias.data = l.bias.data.half()
+            l.bias.data = l.bias.data.float()
+            # l.bias.data = l.bias.data.half()
 
 
 def convert_module_to_f32(l):
